WIP: Add missing fakers for tests
